speech-to-text-app/backend/requirements.txt
flask
flask-cors
flask-socketio
python-dotenv
pydub
speechrecognition
pyjwt
requests
ffmpeg-python
flask
flask_cors
flask_sqlalchemy
requests
pyjwt
werkzeug
